+++ src/test/groovy/com/github/tonybaines/gestalt/BasicBehaviourSpec.groovy
@@ -98,4 +98,21 @@ class BasicBehaviourSpec extends Specification {
     e.cause.message.contains('more than one definition')
   }
 
+  interface Truth {
+    boolean getBool()
+  }
+
+  def "Issue 17: False boolean value is treated as not found"() {
+    when:
+    Properties props = new Properties()
+    props.setProperty('bool', 'false')
+    Truth configInstance = Configurations.definedBy(Truth)
+      .fromProperties(props)
+      .done()
+
+    then:
+    configInstance.getBool() == false
+
+  }
+
 }